+++ packages/@codotype/ui/src/components/configuration_group_input/__tests__/ConfigurationCollectionInput.stories.tsx
@@ -2,19 +2,148 @@ import * as React from "react";
 import { storiesOf } from "@storybook/react";
 import { ConfigurationCollectionInput } from "../ConfigurationCollectionInput";
 import {
-    OptionValueInstance,
     testState,
     ConfigurationGroup,
     OptionValue,
-    buildConfigurationGroupValue,
     buildConfigurationGroupPropertyValue,
     buildTokenPluralization,
+    Codotype,
+    OptionType,
+    PropertyLayoutVariant,
+    StringValueFilter,
+    DataPreviewLayoutVariant,
+    DataPreviewActionType,
+    DataPreviewConstraintType,
 } from "@codotype/core";
 const { LambdaBuilderConfigurationGroup } = testState;
 import { Story } from "@src/components/dev";
 import { ConfigurationInputFormGroup } from "../ConfigurationInputFormGroup";
 
 // // // //
+
+// TODO - move into `test_data`
+// TODO - export StringValueFilter + NumberValueFilter + Validations from @codotype/type (and include constructors as well!)
+const ApiActionConfigurationGroup: ConfigurationGroup = new Codotype.ConfigurationGroup(
+    {
+        label: "API Actions",
+        identifier: "api_actions",
+        description: "Define individual REST API actions for your API",
+        properties: [
+            new Codotype.ConfigurationGroupProperty({
+                label: "Actions",
+                identifier: "actions",
+                type: OptionType.COLLECTION,
+                dataPreview: {
+                    rules: [
+                        {
+                            constraint: {
+                                dataProperty: "function_name",
+                                type: DataPreviewConstraintType.equals,
+                                value: "",
+                            },
+                            action: {
+                                type: DataPreviewActionType.literal,
+                                template: "",
+                            },
+                        },
+                        {
+                            constraint: {
+                                dataProperty: "route",
+                                type: DataPreviewConstraintType.equals,
+                                value: "",
+                            },
+                            action: {
+                                type: DataPreviewActionType.literal,
+                                template: "",
+                            },
+                        },
+                        {
+                            constraint: {
+                                dataProperty: "scope",
+                                type: DataPreviewConstraintType.equals,
+                                value: "COLLECTION",
+                            },
+                            action: {
+                                type: DataPreviewActionType.stringTemplate,
+                                template:
+                                    "{{data.verb}} /api/schema-scope/{{data.route}} -> {{data.function_name}}",
+                            },
+                        },
+                        {
+                            constraint: {
+                                dataProperty: "scope",
+                                type: DataPreviewConstraintType.equals,
+                                value: "MODEL",
+                            },
+                            action: {
+                                type: DataPreviewActionType.stringTemplate,
+                                template:
+                                    "{{data.verb}} /api/schema-scope/:id/{{data.route}} -> {{data.function_name}}",
+                            },
+                        },
+                    ],
+                    variant: DataPreviewLayoutVariant.CODE_DARK,
+                },
+                properties: [
+                    new Codotype.ConfigurationGroupProperty({
+                        label: "Verb",
+                        identifier: "verb",
+                        description: "Verify",
+                        defaultValue: "GET",
+                        type: OptionType.DROPDOWN,
+                        layoutVariant: PropertyLayoutVariant.COL_6,
+                        dropdownOptions: [
+                            { value: "GET", label: "GET" },
+                            { value: "POST", label: "POST" },
+                            { value: "PUT", label: "PUT" },
+                            { value: "DELETE", label: "DELETE" },
+                        ],
+                    }),
+                    new Codotype.ConfigurationGroupProperty({
+                        label: "Route",
+                        identifier: "route",
+                        description: "Route",
+                        defaultValue: "verify",
+                        type: OptionType.STRING,
+                        layoutVariant: PropertyLayoutVariant.COL_6,
+                        filters: [
+                            StringValueFilter.nonumbers,
+                            StringValueFilter.trimwhitespace,
+                            StringValueFilter.removewhitespace,
+                        ],
+                    }),
+                    new Codotype.ConfigurationGroupProperty({
+                        label: "Function Name",
+                        identifier: "function_name",
+                        description: "function_name",
+                        defaultValue: "verify",
+                        layoutVariant: PropertyLayoutVariant.COL_6,
+                        type: OptionType.STRING,
+                        filters: [
+                            StringValueFilter.camelcase,
+                            StringValueFilter.nonumbers,
+                            StringValueFilter.nosymbols,
+                            StringValueFilter.trimwhitespace,
+                            StringValueFilter.removewhitespace,
+                        ],
+                    }),
+                    new Codotype.ConfigurationGroupProperty({
+                        label: "Scope",
+                        identifier: "scope",
+                        description: "scope",
+                        defaultValue: "COLLECTION",
+                        layoutVariant: PropertyLayoutVariant.COL_6,
+                        type: OptionType.DROPDOWN,
+                        dropdownOptions: [
+                            { value: "COLLECTION", label: "Collection" },
+                            { value: "MODEL", label: "Model" },
+                        ],
+                    }),
+                ],
+            }),
+        ],
+    },
+);
 
 const stories: [string, ConfigurationGroup, OptionValue][] = [
     [
@@ -29,6 +158,7 @@ const stories: [string, ConfigurationGroup, OptionValue][] = [
         LambdaBuilderConfigurationGroup,
         [{ id: "123", lambdaName: "Foobar", language: "typescript" }],
     ],
+    ["API Actions", ApiActionConfigurationGroup, []],
 ];
 
 const storyCollection = storiesOf(
@@ -45,19 +175,16 @@ stories.forEach(story => {
         // TODO - fix this, including extra value by default
         console.log("DEFAULT VALUE");
         console.log(
-            buildConfigurationGroupPropertyValue(
-                LambdaBuilderConfigurationGroup.properties[0],
-            ),
+            buildConfigurationGroupPropertyValue(story[1].properties[0]),
         );
 
         return (
             <Story>
-                <ConfigurationInputFormGroup
-                    property={LambdaBuilderConfigurationGroup.properties[0]}
-                >
+                <ConfigurationInputFormGroup property={story[1].properties[0]}>
                     <ConfigurationCollectionInput
                         identifiers={buildTokenPluralization("Lambda")}
                         properties={story[1].properties[0].properties}
+                        dataPreview={story[1].properties[0].dataPreview}
                         value={value}
                         onChange={(updatedVal: OptionValue) => {
                             setValue(updatedVal);