+++ packages/@codotype/core/CHANGELOG.md
@@ -1,3 +1,13 @@
+### 0.8.0-canary.5
+
+- Export `DataPreview` interfaces + functions
+- Added `GeneratorMeta.exampleProjects`
+
+### 0.8.0-canary.4
+
+- Added `DataPreview` interfaces + functions
+- Added `ConfigurationGroupProperty.dataPreview`
+
 ### 0.8.0-canary.3
 
 - Export `PropertyFilter` + `PropertyValidation` interfaces + enums